реализован Redis через SOLID, импорты в соответствии с pep
